feat: refactor recommendation handling into new controller
- Moved recommendation logic from ModalManagerController to RecommendationController.
- Updated routing for recommendation modal in the HTML template.
- Cleaned up unused imports and code in the original controller.
